Final Merge

+++ JEDI-FLIPFIT-GROUP-A-DROPWIZARD/src/main/java/com/flipkart/restcontroller/GymFlipFitAdminController.java
@@ -22,7 +22,7 @@ public class GymFlipFitAdminController {
 
     @GET
     @Path("/login")
-    public Response customerLogin(@QueryParam("email") String email, @QueryParam("password") String password) {
+    public Response adminLogin(@QueryParam("email") String email, @QueryParam("password") String password) {
         // http://localhost:8080/admin/login?email=admin@example.com&password=password123
         userId = userService.authenticateUser(email, password, 3);
         System.out.println("USER ID: " + userId);
@@ -39,6 +39,9 @@ public class GymFlipFitAdminController {
     @Timed
     public Response getViewAllGymOwners() {
         // http://localhost:8080/admin/view-gym-owners
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try{
             List<FlipFitGymOwner> gymOwnerList = adminService.viewAllGymOwners();
             return Response.ok(gymOwnerList).build();
@@ -52,11 +55,14 @@ public class GymFlipFitAdminController {
     @Timed
     public Response viewGymOwnerDetails(@QueryParam("ownerId") int ownerId) {
         // http://localhost:8080/admin/view-gym-owner-details?ownerId=1
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try {
             List<FlipFitGymOwner> gymOwnerList = adminService.viewGymOwnerDetails(ownerId);
             return Response.ok(gymOwnerList).build();
         } catch (Exception e){
-            return Response.status(400).entity("Invalid Gym Owner details").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
 
@@ -65,11 +71,14 @@ public class GymFlipFitAdminController {
     @Timed
     public Response viewGymOwnerRequests() {
         // http://localhost:8080/admin/view-gym-owner-requests
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try {
             List<FlipFitGymOwner> gymOwnerList = adminService.viewGymOwnerRequests();
             return Response.ok(gymOwnerList).build();
         } catch (Exception e) {
-            return Response.status(400).entity("Invalid Gym Owner Requests").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
 
@@ -78,11 +87,14 @@ public class GymFlipFitAdminController {
     @Timed
     public Response approveGymOwnerRequests(@QueryParam("ownerId") int ownerId) {
         // http://localhost:8080/admin/approve-gym-owner-requests?ownerId=1
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try {
             adminService.approveGymOwnerRequests(ownerId);
-            return Response.ok("Approved the gym owner requests with Id").build();
+            return Response.ok("Approved the gym owner requests with Id "+ownerId).build();
         } catch (Exception e) {
-            return Response.status(400).entity("Invalid Gym Owner Requests").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
 
@@ -91,11 +103,14 @@ public class GymFlipFitAdminController {
     @Timed
     public Response removeGymOwner(@QueryParam("ownerId") int ownerId) {
         // http://localhost:8080/admin/remove-gym-owner?ownerId=1
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try {
             adminService.removeGymOwner(ownerId);
-            return Response.ok("Removed the gym owner with Id").build();
+            return Response.ok("Removed the gym owner with Id "+ownerId).build();
         } catch (Exception e) {
-            return Response.status(400).entity("Invalid Gym Owner").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
     }
 
@@ -104,11 +119,25 @@ public class GymFlipFitAdminController {
     @Timed
     public Response cancelRequest(@QueryParam("ownerId") int ownerId) {
         // http://localhost:8080/admin/cancel-request?ownerId=1
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
         try {
             adminService.cancelRequest(ownerId);
             return Response.ok("Request cancelled").build();
         } catch (Exception e) {
-            return Response.status(400).entity("Invalid Gym Owner").build();
+            return Response.status(Response.Status.UNAUTHORIZED).entity(e.getMessage()).build();
         }
+    }
+
+    @GET
+    @Path("/logout")
+    public Response adminLogOut() {
+        if (userId <= 0) {
+            return Response.status(Response.Status.UNAUTHORIZED).entity("You are not an authorized user!").build();
+        }
+        userId = -1;
+        return Response.ok("Logged out successfully").build();
+
     }
 }
